requirements.txt
aiocron==1.8
aiohttp==3.9.5
checksumdir==1.2.0
dacite==1.8.1
docker==7.0.0
fastapi==0.111.0
pydantic==2.7.1
PyQt5==5.15.10
python-qbittorrent==0.4.3
setuptools==69.5.1
torrentool==1.2.0
uvicorn=0.29.0
pandas==2.2.2
pydantic==2.7.1
pytest==8.2.0
PyYAML==6.0.1
scikit-learn==1.4.2
SQLAlchemy==2.0.30